[NEST-FE-26]Extract TagsSelector as a smaller component from CreatePost component
